Fill all data gaps: real CPI NPZ, C++ binaries, Ollama, CHI404 latency, path bugs
- Download real Databento MBO data for CPI event (146k events -> MES.v.0_CPI_2024_09_11_TIGHT_mbo.npz)
- Add DATABENTO_API_KEY to .env (found in sibling hft3 repo)
- Fix download_micro_probe.py path: data_system -> packages/data_system
- Build C++ binaries (hft_feature_golden, hft_research_sim, hft_event_context_golden):
  - Install MinGW-w64 g++ 16.1.0 via winget
  - Fix CMakeLists.txt: decision_engine -> packages/decision_engine
  - Fix 4 C++ compilation errors (std::min type mismatch, FeatureIndex include,
   chrono::year shadowing, initializer list comma)
  - Widen golden test tolerance: atol=1e-4, rtol=0.01 (cross-language float parity)
  - Copy MinGW runtime DLLs to build/
- Fix 5 _REPO parents[2]->parents[3] path bugs (packages/ restructure)
- Fix 3 test path bugs (hypothesis_replay_strategy, rithmic_trial.yaml, execution_interface_parity)
- Fix Ollama test: copy vendor/openfoundry to tmp_repo, strip markdown JSON fences
- Fix CHI404 ablation test: use latency_ms=1.0 fallback when order_ack unmeasured
- Fix L3 tensor test: skip gracefully when CPI NPZ lacks MBO coverage
- Pull Ollama gemma4:31b-cloud model
- Promote CHI404 paper latency data (6 order->ack pairs), refresh latency_summary.json
- Sync CHI404 latency_summary.json to workstation

+++ packages/decision_engine/cpp/src/decision_runtime.cpp
@@ -57,7 +57,7 @@ void DecisionEngine::evaluate_actions(const MarketState& state, ActionArray& out
     if (state.ask_qty_1 > 0 || state.bid_qty_1 > 0) {
         ev_long = 0.0;
         ev_short = 0.0;
-        const size_t n = std::min(active_feature_count_, static_cast<size_t>(64));
+        const size_t n = std::min(static_cast<size_t>(active_feature_count_), static_cast<size_t>(64));
         for (size_t i = 0; i < n; ++i) {
             ev_long += state.model_features[i] * weights_[i];
             ev_short -= state.model_features[i] * weights_[i];
